project/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
bcrypt==4.0.1
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.2
httpx==0.25.2
cryptography==41.0.7
python-multipart==0.0.6
pydantic==2.5.0